pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "git-crew"
version = "0.1.0"
requires-python = ">=3.11"
readme = "README.md"
dependencies = [
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-groq>=0.2.0",
    "crewai>=0.80.0",
    "ag2>=0.11.0",
    "typer>=0.12.0",
    "rich>=13.0.0",
    "typing-extensions>=4.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
git-crew = "gitcrew.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["gitcrew*"]
